Special Guest
Added interactive(java at the top of html) special guest
information/panel, they all have the same picture at the moment cause
that was all i could find in share folder with special guest images.
Secondly image is currently formated in html so as to fit in panel cause
of issues when full size. The image i use need to be photoshop as it has
a black background that cant be see and is too big for the panel.

If someone could photoshop that for me that would be good and if you did
the size of the photo should the format size i have set in html. and
could this be done with all special guest photos, thx