about history
I've done some work on the history page to make navigation easier. I've
added a list of the dates at the top which warps to that date on the
page. I also added back to top button links at the end of each section.
I fixed a bug that would refresh the page when clicking anywhere in the
content.
Also I fixed a bug in the footer that would think a link is being
hovered when it wasn't.
Lastly I reduced the number of words required in the message field of
the form to 20 characters.